Hooks for "review published" and "review request published".
`/rb/review-published` checks for `ship_it` property and if present, adds a
`reviewed` label to the corresponding PT story or resolves a remote link in the
correspoding JIRA issue (according to the `branch` field).

`/rb/review-request-published` checks for updates to the review request and
adds a comment to the correspoding PT story or adds a remote link to the
corresponding JIRA issue (according to the `branch` field).