Transition issue to "reviewed" state if all reviews are approved.
We check JIRA issue's linked remote issues (which is how we link review
requests) to see if all reviews have been approved. We don't check RB because
RB's API is silly and does not allow searching by the `bug` field so finding
the review requests for a certain JIRA issue would be hard.

We don't check the initial state of the JIRA issue to be transitioned. We
assume that's JIRA's responsibility; the "mark as reviewed" transition is only
legal from the "implemented" state.

Change-Id: da725c0b13
Story-Id: SF-72